Lint.
git-svn-id: http://svn.drobilla.net/lad/trunk/raul@4388 a436a847-0d15-0410-975c-d299462d15a1

+++ test/queue_test.cpp
@@ -95,10 +95,10 @@ dump_data()
 int
 main()
 {
-	unsigned long total_processed = 0;
+	size_t total_processed = 0;
 
 	cout << "Testing size" << endl;
-	for (unsigned i=0; i < queue.capacity(); ++i) {
+	for (size_t i=0; i < queue.capacity(); ++i) {
 		queue.push(i);
 		if (i == queue.capacity()-1) {
 			if (!queue.full()) {
@@ -115,7 +115,7 @@ main()
 		}
 	}
 
-	for (unsigned i = 0; i < queue.capacity(); ++i)
+	for (size_t i = 0; i < queue.capacity(); ++i)
 		queue.pop();
 
 	if (!queue.empty()) {
